[REF] point_of_sale: merge pos_epson_printer in point_of_sale
We simplify the logic by merging pos epson printer inside of point of sale.

Upgrade: odoo/upgrade#8223
