+++ addons/point_of_sale/static/src/app/components/epos_templates.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <templates id="template" xml:space="preserve">
-    <t t-name="pos_epson_printer.ePOSLayout">
+    <t t-name="point_of_sale.ePOSLayout">
         <s:Envelope xmlns:s="http://schemas.xmlsoap.org/soap/envelope/">
             <s:Body>
                 <epos-print xmlns="http://www.epson-pos.com/schemas/2011/03/epos-print" />
