+++ addons/point_of_sale/views/pos_printer_view.xml
@@ -9,6 +9,7 @@
                         <field name="company_id" invisible="1" />
                         <field name="name" />
                         <field name="printer_type" widget="radio"/>
+                        <field name="epson_printer_ip" invisible="printer_type != 'epson_epos'"/>
                         <field name="proxy_ip" invisible="printer_type != 'iot'"/>
                         <field name="product_categories_ids" />
                     </group>
