+++ addons/point_of_sale/views/res_config_settings_views.xml
@@ -386,6 +386,16 @@
                             <block title="Connected Devices" id="pos_connected_devices_section">
                                 <setting id="pos_other_devices" string="ePos Printer" help="Connect device to your PoS without an IoT Box" documentation="/applications/sales/point_of_sale/configuration/epos_ssc.html">
                                     <field name="pos_other_devices"/>
+                                    <div class="content-group" invisible="not pos_other_devices">
+                                        <field name="pos_epson_printer_ip" placeholder="Epson Receipt Printer IP Address" />
+                                        <div class="row" invisible="pos_epson_printer_ip in [False, '']">
+                                            <label string="Cashdrawer" for="pos_iface_cashdrawer" class="col-lg-3 o_light_label"/>
+                                            <field name="pos_iface_cashdrawer"/>
+                                        </div>
+                                    </div>
+                                    <div role="alert" class="alert alert-warning" invisible="not pos_iface_print_via_proxy or not pos_other_devices or pos_epson_printer_ip in [False, '']">
+                                        The Epson receipt printer will be used instead of the receipt printer connected to the IoT Box.
+                                    </div>
                                 </setting>
                                 <setting id="customer_display" string="Customer Display" help="Set the background image of your customer display" documentation="/applications/sales/point_of_sale/shop/customer_display.html" >
                                     <label for="pos_customer_display_bg_img" />
