--- addons/pos_epson_printer/i18n/nl.po
@@ -1,161 +1,0 @@
-# Translation of Odoo Server.
-# This file contains the translation of the following modules:
-# 	* pos_epson_printer
-#
-# Translators:
-# Martin Trigaux, 2022
-# Erwin van der Ploeg <erwin@odooexperts.nl>, 2022
-# Jolien De Paepe, 2023
-#
-msgid ""
-msgstr ""
-"Project-Id-Version: Odoo Server 16.0\n"
-"Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2023-05-16 13:48+0000\n"
-"PO-Revision-Date: 2022-09-22 05:54+0000\n"
-"Last-Translator: Jolien De Paepe, 2023\n"
-"Language-Team: Dutch (https://app.transifex.com/odoo/teams/41243/nl/)\n"
-"Language: nl\n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: \n"
-"Plural-Forms: nplurals=2; plural=(n != 1);\n"
-
-#. module: pos_epson_printer
-#: model_terms:ir.ui.view,arch_db:pos_epson_printer.pos_iot_config_view_form
-#: model_terms:ir.ui.view,arch_db:pos_epson_printer.res_config_settings_view_form
-msgid "Cashdrawer"
-msgstr "Geldlade"
-
-#. module: pos_epson_printer
-#. odoo-javascript
-#: code:addons/pos_epson_printer/static/src/js/epson_printer.js:0
-msgid "Check on the printer configuration for the 'Device ID' setting. It should be set to: "
-msgstr "Controleer in de printerconfiguratie de instelling 'ID Apparaat'. Deze moet zijn ingesteld op:"
-
-#. module: pos_epson_printer
-#: model:ir.model,name:pos_epson_printer.model_res_config_settings
-msgid "Config Settings"
-msgstr "Configuratie instellingen"
-
-#. module: pos_epson_printer
-#. odoo-javascript
-#: code:addons/pos_epson_printer/static/src/js/epson_printer.js:0
-msgid "Connection to the printer failed"
-msgstr "Verbinding met de printer mislukt"
-
-#. module: pos_epson_printer
-#: model:ir.model.fields,field_description:pos_epson_printer.field_pos_config__epson_printer_ip
-msgid "Epson Printer IP"
-msgstr "Epson Printer IP"
-
-#. module: pos_epson_printer
-#: model:ir.model.fields,field_description:pos_epson_printer.field_pos_printer__epson_printer_ip
-msgid "Epson Printer IP Address"
-msgstr ""
-
-#. module: pos_epson_printer
-#. odoo-python
-#: code:addons/pos_epson_printer/models/pos_printer.py:0
-msgid "Epson Printer IP Address cannot be empty."
-msgstr ""
-
-#. module: pos_epson_printer
-#: model_terms:ir.ui.view,arch_db:pos_epson_printer.pos_iot_config_view_form
-#: model_terms:ir.ui.view,arch_db:pos_epson_printer.res_config_settings_view_form
-msgid "Epson Receipt Printer IP Address"
-msgstr "Epson ticketprinter IP adres"
-
-#. module: pos_epson_printer
-#. odoo-javascript
-#: code:addons/pos_epson_printer/static/src/js/epson_printer.js:0
-msgid "If you are on a secure server (HTTPS) please make sure you manually accepted the certificate by accessing %s"
-msgstr "Als je je op een beveiligde server (HTTPS) bevindt, zorg er dan voor dat je het certificaat handmatig hebt geaccepteerd door toegang te krijgen tot %s"
-
-#. module: pos_epson_printer
-#: model:ir.model.fields,help:pos_epson_printer.field_pos_config__epson_printer_ip
-#: model:ir.model.fields,help:pos_epson_printer.field_pos_printer__epson_printer_ip
-msgid "Local IP address of an Epson receipt printer."
-msgstr "Lokaal IP adres van een Epson ticketprinter."
-
-#. module: pos_epson_printer
-#. odoo-javascript
-#: code:addons/pos_epson_printer/static/src/js/epson_printer.js:0
-msgid "No paper was detected by the printer"
-msgstr "De printer heeft geen papier gedetecteerd"
-
-#. module: pos_epson_printer
-#. odoo-javascript
-#: code:addons/pos_epson_printer/static/src/js/epson_printer.js:0
-msgid "Please check if the printer has enough paper and is ready to print."
-msgstr "Controleer of de printer voldoende papier heeft en klaar is om af te drukken."
-
-#. module: pos_epson_printer
-#. odoo-javascript
-#: code:addons/pos_epson_printer/static/src/js/epson_printer.js:0
-msgid ""
-"Please check if the printer is still connected. \n"
-"Some browsers don't allow HTTP calls from websites to devices in the network (for security reasons). If it is the case, you will need to follow Odoo's documentation for 'Self-signed certificate for ePOS printers' and 'Secure connection (HTTPS)' to solve the issue"
-msgstr ""
-"Controleer of de printer nog is aangesloten.\n"
-"Sommige browsers staan geen HTTP-oproepen van websites naar apparaten in het netwerk toe (om veiligheidsredenen). Als dit het geval is, moet je de Odoo-documentatie voor 'Zelfondertekend certificaat voor ePOS-printers' en 'Beveiligde verbinding (HTTPS)' volgen om het probleem op te lossen"
-
-#. module: pos_epson_printer
-#: model:ir.model,name:pos_epson_printer.model_pos_config
-msgid "Point of Sale Configuration"
-msgstr "Kassa instellingen"
-
-#. module: pos_epson_printer
-#: model:ir.model,name:pos_epson_printer.model_pos_printer
-msgid "Point of Sale Printer"
-msgstr ""
-
-#. module: pos_epson_printer
-#: model:ir.model,name:pos_epson_printer.model_pos_session
-msgid "Point of Sale Session"
-msgstr ""
-
-#. module: pos_epson_printer
-#: model:ir.model.fields,field_description:pos_epson_printer.field_res_config_settings__pos_epson_printer_ip
-msgid "Pos Epson Printer Ip"
-msgstr "Kassa Epson Printer Ip"
-
-#. module: pos_epson_printer
-#: model:ir.model.fields,field_description:pos_epson_printer.field_pos_printer__printer_type
-msgid "Printer Type"
-msgstr ""
-
-#. module: pos_epson_printer
-#. odoo-javascript
-#: code:addons/pos_epson_printer/static/src/js/epson_printer.js:0
-msgid "Printing failed"
-msgstr "Afdrukken mislukt"
-
-#. module: pos_epson_printer
-#: model_terms:ir.ui.view,arch_db:pos_epson_printer.pos_iot_config_view_form
-#: model_terms:ir.ui.view,arch_db:pos_epson_printer.res_config_settings_view_form
-msgid "The Epson receipt printer will be used instead of the receipt printer connected to the IoT Box."
-msgstr "De Epson ticketprinter wordt gebruikt in plaats van de ticketprinter gekoppeld aan de IoT box."
-
-#. module: pos_epson_printer
-#. odoo-javascript
-#: code:addons/pos_epson_printer/static/src/js/epson_printer.js:0
-msgid "The following error code was given by the printer:"
-msgstr "De printer gaf de volgende foutcode:"
-
-#. module: pos_epson_printer
-#. odoo-javascript
-#: code:addons/pos_epson_printer/static/src/js/epson_printer.js:0
-msgid "The printer was successfully reached, but it wasn't able to print."
-msgstr "De printer werd met succes bereikt, maar kon niet afdrukken."
-
-#. module: pos_epson_printer
-#. odoo-javascript
-#: code:addons/pos_epson_printer/static/src/js/epson_printer.js:0
-msgid "To find more details on the error reason, please search online for:"
-msgstr "Voor meer details over de reden van de fout kan je online zoeken naar:"
-
-#. module: pos_epson_printer
-#: model:ir.model.fields.selection,name:pos_epson_printer.selection__pos_printer__printer_type__epson_epos
-msgid "Use an Epson printer"
-msgstr ""
