--- addons/pos_epson_printer/i18n/zh_TW.po
@@ -1,158 +1,0 @@
-# Translation of Odoo Server.
-# This file contains the translation of the following modules:
-# 	* pos_epson_printer
-# 
-# Translators:
-# Wil Odoo, 2025
-# 
-msgid ""
-msgstr ""
-"Project-Id-Version: Odoo Server saas~18.2\n"
-"Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2025-02-20 10:02+0000\n"
-"PO-Revision-Date: 2025-05-06 12:03+0000\n"
-"Last-Translator: Wil Odoo, 2025\n"
-"Language-Team: Chinese (Taiwan) (https://app.transifex.com/odoo/teams/41243/zh_TW/)\n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: \n"
-"Language: zh_TW\n"
-"Plural-Forms: nplurals=1; plural=0;\n"
-
-#. module: pos_epson_printer
-#: model_terms:ir.ui.view,arch_db:pos_epson_printer.pos_iot_config_view_form
-#: model_terms:ir.ui.view,arch_db:pos_epson_printer.res_config_settings_view_form
-msgid "Cashdrawer"
-msgstr "錢箱"
-
-#. module: pos_epson_printer
-#. odoo-javascript
-#: code:addons/pos_epson_printer/static/src/app/utils/payment/epson_printer.js:0
-msgid ""
-"Check on the printer configuration for the 'Device ID' setting. It should be"
-" set to: "
-msgstr "請檢查打印機配置中的\"設備 ID\",應設置為："
-
-#. module: pos_epson_printer
-#: model:ir.model,name:pos_epson_printer.model_res_config_settings
-msgid "Config Settings"
-msgstr "配置設定"
-
-#. module: pos_epson_printer
-#: model:ir.model.fields,field_description:pos_epson_printer.field_pos_config__display_name
-#: model:ir.model.fields,field_description:pos_epson_printer.field_pos_printer__display_name
-#: model:ir.model.fields,field_description:pos_epson_printer.field_res_config_settings__display_name
-msgid "Display Name"
-msgstr "顯示名稱"
-
-#. module: pos_epson_printer
-#: model:ir.model.fields,field_description:pos_epson_printer.field_pos_config__epson_printer_ip
-msgid "Epson Printer IP"
-msgstr "Epson 印表機 IP"
-
-#. module: pos_epson_printer
-#: model:ir.model.fields,field_description:pos_epson_printer.field_pos_printer__epson_printer_ip
-msgid "Epson Printer IP Address"
-msgstr "Epson打印機IP地址"
-
-#. module: pos_epson_printer
-#. odoo-python
-#: code:addons/pos_epson_printer/models/pos_printer.py:0
-msgid "Epson Printer IP Address cannot be empty."
-msgstr "Epson 打印機IP地址不能為空。"
-
-#. module: pos_epson_printer
-#: model_terms:ir.ui.view,arch_db:pos_epson_printer.pos_iot_config_view_form
-#: model_terms:ir.ui.view,arch_db:pos_epson_printer.res_config_settings_view_form
-msgid "Epson Receipt Printer IP Address"
-msgstr "Epson 收據印表機 IP 地址"
-
-#. module: pos_epson_printer
-#: model:ir.model.fields,field_description:pos_epson_printer.field_pos_config__id
-#: model:ir.model.fields,field_description:pos_epson_printer.field_pos_printer__id
-#: model:ir.model.fields,field_description:pos_epson_printer.field_res_config_settings__id
-msgid "ID"
-msgstr "識別號"
-
-#. module: pos_epson_printer
-#. odoo-javascript
-#: code:addons/pos_epson_printer/static/src/app/utils/payment/epson_printer.js:0
-msgid ""
-"If you are on a secure server (HTTPS) please make sure you manually accepted"
-" the certificate by accessing %s. "
-msgstr "如果您使用的是安全服務器（HTTPS）,請確保您通過存取 %s 手動接受了證書。"
-
-#. module: pos_epson_printer
-#: model:ir.model.fields,help:pos_epson_printer.field_pos_config__epson_printer_ip
-#: model:ir.model.fields,help:pos_epson_printer.field_pos_printer__epson_printer_ip
-msgid "Local IP address of an Epson receipt printer."
-msgstr "Epson 收據印表機的本地 IP 地址."
-
-#. module: pos_epson_printer
-#. odoo-javascript
-#: code:addons/pos_epson_printer/static/src/app/utils/payment/epson_printer.js:0
-msgid "No paper was detected by the printer"
-msgstr "打印機偵測不到有紙張"
-
-#. module: pos_epson_printer
-#. odoo-javascript
-#: code:addons/pos_epson_printer/static/src/app/utils/payment/epson_printer.js:0
-msgid "Please check if the printer has enough paper and is ready to print."
-msgstr "請檢查印表機是否有足夠的紙張並準備好列印。"
-
-#. module: pos_epson_printer
-#: model:ir.model,name:pos_epson_printer.model_pos_config
-msgid "Point of Sale Configuration"
-msgstr "POS設定"
-
-#. module: pos_epson_printer
-#: model:ir.model,name:pos_epson_printer.model_pos_printer
-msgid "Point of Sale Printer"
-msgstr "POS打印機"
-
-#. module: pos_epson_printer
-#: model:ir.model.fields,field_description:pos_epson_printer.field_res_config_settings__pos_epson_printer_ip
-msgid "Pos Epson Printer Ip"
-msgstr "Pos Epson打印機IP地址"
-
-#. module: pos_epson_printer
-#: model:ir.model.fields,field_description:pos_epson_printer.field_pos_printer__printer_type
-msgid "Printer Type"
-msgstr "印表機類型"
-
-#. module: pos_epson_printer
-#. odoo-javascript
-#: code:addons/pos_epson_printer/static/src/app/utils/payment/epson_printer.js:0
-msgid "Printing failed"
-msgstr "列印失敗"
-
-#. module: pos_epson_printer
-#: model_terms:ir.ui.view,arch_db:pos_epson_printer.pos_iot_config_view_form
-#: model_terms:ir.ui.view,arch_db:pos_epson_printer.res_config_settings_view_form
-msgid ""
-"The Epson receipt printer will be used instead of the receipt printer "
-"connected to the IoT Box."
-msgstr "將使用 Epson 收據印表機替换 IoT 收據印表機."
-
-#. module: pos_epson_printer
-#. odoo-javascript
-#: code:addons/pos_epson_printer/static/src/app/utils/payment/epson_printer.js:0
-msgid "The following error code was given by the printer:"
-msgstr "打印機提供了以下錯誤代碼:"
-
-#. module: pos_epson_printer
-#. odoo-javascript
-#: code:addons/pos_epson_printer/static/src/app/utils/payment/epson_printer.js:0
-msgid "The printer was successfully reached, but it wasn't able to print."
-msgstr "已成功連接打印機，但無法列印。"
-
-#. module: pos_epson_printer
-#. odoo-javascript
-#: code:addons/pos_epson_printer/static/src/app/utils/payment/epson_printer.js:0
-msgid "To find more details on the error reason, please search online for:"
-msgstr "若想了解錯誤原因的更多詳情，請在網上搜尋："
-
-#. module: pos_epson_printer
-#: model:ir.model.fields.selection,name:pos_epson_printer.selection__pos_printer__printer_type__epson_epos
-msgid "Use an Epson printer"
-msgstr "使用 Epson 印表機"
