--- addons/pos_epson_printer/models/__init__.py
@@ -1,6 +1,0 @@
-# -*- coding: utf-8 -*-
-# Part of Odoo. See LICENSE file for full copyright and licensing details.
-
-from . import pos_config
-from . import res_config_settings
-from . import pos_printer
