--- addons/pos_epson_printer/models/pos_config.py
@@ -1,10 +1,0 @@
-# -*- coding: utf-8 -*-
-# Part of Odoo. See LICENSE file for full copyright and licensing details.
-
-from odoo import fields, models
-
-
-class PosConfig(models.Model):
-    _inherit = 'pos.config'
-
-    epson_printer_ip = fields.Char(string='Epson Printer IP', help="Local IP address of an Epson receipt printer.")
