--- addons/pos_epson_printer/views/pos_config_views.xml
@@ -1,26 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-    <record id="pos_iot_config_view_form" model="ir.ui.view">
-        <field name="name">pos.iot.config.form.view</field>
-        <field name="model">pos.config</field>
-        <field name="inherit_id" ref="point_of_sale.pos_config_view_form"/>
-        <field name="arch" type="xml">
-            <xpath expr="//setting[@id='other_devices']" position="inside">
-                <div class="content-group" invisible="not other_devices">
-                    <field name="epson_printer_ip" placeholder="Epson Receipt Printer IP Address" />
-                    <div class="row" invisible="epson_printer_ip in [False, '']">
-                        <div class="col-lg-5 o_light_label">
-                            <label string="Cashdrawer" for="iface_cashdrawer"/>
-                        </div>
-                        <div class="col-lg-7">
-                            <field name="iface_cashdrawer"/>
-                        </div>
-                    </div>
-                </div>
-                <div role="alert" class="alert alert-warning" invisible="not iface_print_via_proxy or not other_devices or epson_printer_ip in [False, '']">
-                    The Epson receipt printer will be used instead of the receipt printer connected to the IoT Box.
-                </div>
-            </xpath>
-        </field>
-    </record>
-</odoo>
