--- addons/pos_epson_printer/views/pos_printer_views.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-    <record id="view_pos_printer_form" model="ir.ui.view">
-        <field name="name">pos.iot.config.form.view</field>
-        <field name="model">pos.printer</field>
-        <field name="inherit_id" ref="point_of_sale.view_pos_printer_form"/>
-        <field name="arch" type="xml">
-            <field name="printer_type" position="after">
-                <field name="epson_printer_ip" invisible="printer_type != 'epson_epos'"/>
-            </field>
-        </field>
-    </record>
-</odoo>
