+++ addons/pos_self_order/__manifest__.py
@@ -63,15 +63,15 @@
             "point_of_sale/static/src/app/components/order_display/*",
             "point_of_sale/static/src/app/components/orderline/*",
             "point_of_sale/static/src/app/components/centered_icon/*",
+            "point_of_sale/static/src/app/components/epos_templates.xml",
             "point_of_sale/static/src/css/pos_receipts.css",
             "point_of_sale/static/src/app/screens/receipt_screen/receipt/**/*",
             "pos_self_order/static/src/overrides/components/receipt_header/*",
-            "point_of_sale/static/src/app/utils/printer/base_printer.js",
+            "point_of_sale/static/src/app/utils/printer/*",
             "point_of_sale/static/src/app/services/printer_service.js",
             'point_of_sale/static/src/app/utils/html-to-image.js',
             "point_of_sale/static/src/app/services/render_service.js",
             "pos_self_order/static/src/app/**/*",
-            "point_of_sale/static/src/app/utils/printer/hw_printer.js",
             "web/static/src/core/utils/render.js",
             "pos_self_order/static/src/app/store/order_change_receipt_template.xml",
             "account/static/src/helpers/*.js",
