+++ addons/web/tooling/_jsconfig.json
@@ -42,7 +42,6 @@
             "@l10n_sa_pos/*": ["addons/l10n_sa_pos/static/src/*"],
             "@pos_adyen/*": ["addons/pos_adyen/static/src/*"],
             "@pos_discount/*": ["addons/pos_discount/static/src/*"],
-            "@pos_epson_printer/*": ["addons/pos_epson_printer/static/src/*"],
             "@pos_hr/*": ["addons/pos_hr/static/src/*"],
             "@pos_hr_restaurant/*": ["addons/pos_hr_restaurant/static/src/*"],
             "@pos_loyalty/*": ["addons/pos_loyalty/static/src/*"],
